feat(interfaces): create ICar

+++ src/interfaces/IVehicle.ts
@@ -2,10 +2,10 @@ import { z } from 'zod';
 
 const VehicleZodSchema = z.object({
   model: z.string().min(3),
-  year: z.number().min(1900).max(2022),
+  year: z.number().min(1900).max(2022).int(),
   color: z.string().min(3),
   status: z.boolean().optional(),
-  buyValue: z.number(),
+  buyValue: z.number().int(),
 });
 
 type IVehicle = z.infer<typeof VehicleZodSchema>;
